// api/seller/client/models/PrismaJson_DiscountBoostSale.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { DiscountBoostSaleType } from './DiscountBoostSaleType';
export type PrismaJson_DiscountBoostSale = Array<{
  quantity?: number;
  value: number;
  productId: number;
  type: DiscountBoostSaleType;
}>;